disable schema-based completion when XML extension is enabled
